event-trapping/requirements.txt
locustio==0.7.5
pika==0.10.0
